# Removing the dependency on Microsoft.IO.RecyclableStream and improving serializer performance