--- a/src/ZiggyCreatures.FusionCache.Serialization.NeueccMessagePack/FusionCacheNeueccMessagePackSerializer.cs
+++ b/src/ZiggyCreatures.FusionCache.Serialization.NeueccMessagePack/FusionCacheNeueccMessagePackSerializer.cs
@@ -1,7 +1,12 @@
-﻿using System.Threading;
+﻿using System;
+using System.Collections.Concurrent;
+using System.Runtime.CompilerServices;
+using System.Threading;
 using System.Threading.Tasks;
 using MessagePack;
 using MessagePack.Resolvers;
+
+using ZiggyCreatures.Caching.Fusion.Internals;
 
 namespace ZiggyCreatures.Caching.Fusion.Serialization.NeueccMessagePack;
 
@@ -11,7 +16,7 @@ namespace ZiggyCreatures.Caching.Fusion.Serialization.NeueccMessagePack;
 public class FusionCacheNeueccMessagePackSerializer
 	: IFusionCacheSerializer
 {
-	/// <summary>
+/// <summary>
 	/// The options class for the <see cref="FusionCacheNeueccMessagePackSerializer"/> class.
 	/// </summary>
 	public class Options
@@ -45,18 +50,31 @@ public class FusionCacheNeueccMessagePackSerializer
 	private readonly MessagePackSerializerOptions? _serializerOptions;
 
 	/// <inheritdoc />
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public byte[] Serialize<T>(T? obj)
 	{
-		return MessagePackSerializer.Serialize<T?>(obj, _serializerOptions);
+		var writer = ArrayPoolBufferWriter.Rent();
+
+		try
+		{
+			MessagePackSerializer.Serialize(writer, obj, _serializerOptions);
+			return writer.ToArray();
+		}
+		finally
+		{
+			ArrayPoolBufferWriter.Return(writer);
+		}
 	}
 
 	/// <inheritdoc />
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public T? Deserialize<T>(byte[] data)
 	{
-		return MessagePackSerializer.Deserialize<T?>(data, _serializerOptions);
+		return MessagePackSerializer.Deserialize<T?>(data.AsMemory(), _serializerOptions);
 	}
 
 	/// <inheritdoc />
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public ValueTask<byte[]> SerializeAsync<T>(T? obj, CancellationToken token = default)
 	{
 		// PER @neuecc 'S SUGGESTION: AVOID AWAITING ON A MEMORY STREAM
@@ -64,6 +82,7 @@ public class FusionCacheNeueccMessagePackSerializer
 	}
 
 	/// <inheritdoc />
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public ValueTask<T?> DeserializeAsync<T>(byte[] data, CancellationToken token = default)
 	{
 		// PER @neuecc 'S SUGGESTION: AVOID AWAITING ON A MEMORY STREAM