--- a/src/ZiggyCreatures.FusionCache.Serialization.NewtonsoftJson/FusionCacheNewtonsoftJsonSerializer.cs
+++ b/src/ZiggyCreatures.FusionCache.Serialization.NewtonsoftJson/FusionCacheNewtonsoftJsonSerializer.cs
@@ -1,9 +1,20 @@
-﻿using System.Text;
+﻿using System.Buffers;
+using System.IO;
+using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
 
+using ZiggyCreatures.Caching.Fusion.Internals;
+
 namespace ZiggyCreatures.Caching.Fusion.Serialization.NewtonsoftJson;
+
+internal class JsonArrayPool : IArrayPool<char>
+{
+	internal static JsonArrayPool Shared { get; } = new JsonArrayPool();
+	public char[] Rent(int minimumLength) => ArrayPool<char>.Shared.Rent(minimumLength);
+	public void Return(char[]? array) => ArrayPool<char>.Shared.Return(array);
+}
 
 /// <summary>
 /// An implementation of <see cref="IFusionCacheSerializer"/> which uses the Newtonsoft Json.NET serializer.
@@ -11,6 +22,7 @@ namespace ZiggyCreatures.Caching.Fusion.Serialization.NewtonsoftJson;
 public class FusionCacheNewtonsoftJsonSerializer
 	: IFusionCacheSerializer
 {
+	private readonly JsonSerializer _jsonSerializer;
 	/// <summary>
 	/// The options class for the <see cref="FusionCacheNewtonsoftJsonSerializer"/> class.
 	/// </summary>
@@ -31,6 +43,7 @@ public class FusionCacheNewtonsoftJsonSerializer
 	public FusionCacheNewtonsoftJsonSerializer(JsonSerializerSettings? settings = null)
 	{
 		_serializerSettings = settings;
+		_jsonSerializer = JsonSerializer.Create(settings);
 	}
 
 	/// <summary>
@@ -48,13 +61,23 @@ public class FusionCacheNewtonsoftJsonSerializer
 	/// <inheritdoc />
 	public byte[] Serialize<T>(T? obj)
 	{
-		return _encoding.GetBytes(JsonConvert.SerializeObject(obj, _serializerSettings));
+		using var stream = new ArrayPoolWritableStream();
+		using var writer = new StreamWriter(stream);
+		using JsonTextWriter jsonWriter = new JsonTextWriter(writer);
+		jsonWriter.ArrayPool = JsonArrayPool.Shared;
+		_jsonSerializer.Serialize(jsonWriter, obj);
+		jsonWriter.Flush();
+		return stream.GetBytes();
 	}
 
 	/// <inheritdoc />
 	public T? Deserialize<T>(byte[] data)
 	{
-		return JsonConvert.DeserializeObject<T>(_encoding.GetString(data), _serializerSettings);
+		using var stream = new MemoryStream(data);
+		using var reader = new StreamReader(stream, _encoding);
+		using var jsonReader = new JsonTextReader(reader);
+		jsonReader.ArrayPool = JsonArrayPool.Shared;
+		return _jsonSerializer.Deserialize<T>(jsonReader);
 	}
 
 	/// <inheritdoc />