--- a/src/ZiggyCreatures.FusionCache.Serialization.ProtoBufNet/FusionCacheProtoBufNetSerializer.cs
+++ b/src/ZiggyCreatures.FusionCache.Serialization.ProtoBufNet/FusionCacheProtoBufNetSerializer.cs
@@ -135,5 +135,5 @@ public class FusionCacheProtoBufNetSerializer
 	}
 
 	/// <inheritdoc />
-	public override string ToString() => $"{(_streamManager != null ? "Recyclable" : "")}{GetType().Name}";
+	public override string ToString() => GetType().Name;
 }