@jodydonetti 2024-12-21
Hi @stebet , I run the benchmarks on my machine, and the results seem to be consistent.

Here are the main results.

## CysharpMemoryPack (Serialize)

| What         | Mean        | Error     | StdDev    | P95         | Gen0     | Gen1     | Gen2     | Allocated  |
|--------------|------------:|----------:|----------:|------------:|---------:|---------:|---------:|-----------:|
| Before       |    97.09 us |  1.927 us |  1.892 us |    99.62 us |  30.2734 |  30.2734 |  30.2734 |   94.79 KB |
| After        |    67.01 us |  0.638 us |  0.597 us |    67.79 us |  30.2734 |  30.2734 |  30.2734 |   94.93 KB |
| After+Shared |   109.27 us |  4.862 us | 14.337 us |   119.33 us |  30.2734 |  30.2734 |  30.2734 |   94.93 KB |

## CysharpMemoryPack (Deserialize)

| What         | Mean        | Error     | StdDev    | P95         | Gen0     | Gen1     | Gen2     | Allocated  |
|--------------|------------:|----------:|----------:|------------:|---------:|---------:|---------:|-----------:|
| Before       |    49.72 us |  0.965 us |  1.073 us |    51.49 us |  21.6675 |   7.0190 |        - |  265.68 KB |
| After        |    34.43 us |  0.411 us |  0.385 us |    34.94 us |  21.6675 |   7.0190 |        - |  265.68 KB |
| After+Shared |    34.51 us |  0.588 us |  0.550 us |    35.25 us |  21.6675 |   7.0190 |        - |  265.68 KB |

## NeueccMessagePack (Serialize)

| What         | Mean        | Error     | StdDev    | P95         | Gen0     | Gen1     | Gen2     | Allocated  |
|--------------|------------:|----------:|----------:|------------:|---------:|---------:|---------:|-----------:|
| Before       |   122.96 us |  1.921 us |  1.703 us |   125.93 us |   5.6152 |        - |        - |   71.59 KB |
| After        |    78.74 us |  1.396 us |  1.306 us |    80.67 us |   5.7373 |        - |        - |   71.62 KB |
| After+Shared |    81.04 us |  0.867 us |  0.811 us |    82.07 us |   5.7373 |        - |        - |   71.62 KB |

## NeueccMessagePack (Deserialize)

| What         | Mean        | Error     | StdDev    | P95         | Gen0     | Gen1     | Gen2     | Allocated  |
|--------------|------------:|----------:|----------:|------------:|---------:|---------:|---------:|-----------:|
|              |             |           |           |             |          |          |          |            |
| Before       |   189.92 us |  3.218 us |  3.010 us |   194.60 us |  21.4844 |   6.8359 |        - |  265.68 KB |
| After        |   150.30 us |  1.558 us |  1.381 us |   152.66 us |  21.4844 |   6.8359 |        - |  265.68 KB |
| After+Shared |   149.99 us |  1.415 us |  1.255 us |   151.69 us |  21.4844 |   6.8359 |        - |  265.68 KB |

## SystemTextJson (Serialize)

| What               | Mean        | Error     | StdDev    | P95         | Gen0     | Gen1     | Gen2     | Allocated  |
|--------------------|------------:|----------:|----------:|------------:|---------:|---------:|---------:|-----------:|
| Before             |   383.09 us |  5.302 us |  4.959 us |   390.19 us |  45.4102 |  45.4102 |  45.4102 |  146.61 KB |
| Before+Recyclable  |   331.87 us |  3.728 us |  3.305 us |   336.54 us |  40.0391 |  40.0391 |  40.0391 |  146.85 KB |
| After              |   250.10 us |  4.506 us |  4.215 us |   257.07 us |  41.5039 |  41.5039 |  41.5039 |  146.91 KB |
| After+Shared       |   288.82 us |  1.824 us |  1.523 us |   290.86 us |  45.4102 |  45.4102 |  45.4102 |  146.61 KB |


## SystemTextJson (Deserialize)

| What               | Mean        | Error     | StdDev    | P95         | Gen0     | Gen1     | Gen2     | Allocated  |
|--------------------|------------:|----------:|----------:|------------:|---------:|---------:|---------:|-----------:|
| Before             |   644.87 us | 12.249 us | 11.457 us |   662.22 us |  31.2500 |   9.7656 |        - |  391.67 KB |
| Before+Recyclable  |   634.01 us |  2.830 us |  2.363 us |   638.22 us |  31.2500 |   9.7656 |        - |  391.67 KB |
| After              |   511.61 us |  3.509 us |  3.283 us |   516.57 us |  31.2500 |   9.7656 |        - |  391.67 KB |
| After+Shared       |   530.56 us |  6.669 us |  6.238 us |   536.86 us |  31.2500 |   9.7656 |        - |  391.67 KB |

As you can see for System.Text.Json I also added the version with the `RecyclableMemoryStream`.

I also added a version that used `ArrayPool<byte>.Shared` instead of `ArrayPool<byte>.Create()` for a (theoretically) higher memory reuse, but as you can see it looks like not the case, and actually seems even slightly worse.

I'm thinking that we may see advantages in using the `Shared` in real applications with a lot of allocations (so, more use of the shared pool), but I'm not sure.

I'd say to just go with the `Create()` one based on these results.