@jodydonetti 2024-12-21
Also, I need to decide what to do with the `RecyclableMemoryStream` dependency.

On one hand I can just remove it and everything will be slimmer, but then I'd have to also remove the ctor overload that had a param of such type.
Usually I would mark the ctor as [Obsolete] and just forward the call to the (now automatically optimized) other ctor, but to keep it I need to keep the package ref, otherwise the type itself is not there anymore and it won't compile of course

On the other hand we are talking about V2, which is a major version bump already with some breaking changes, so there's that.

In the end what I don't like much is the developer experience of upgrading and having compile-time errors without any specific error message.
Mmmh, doubts, thoughts.

Pinging @viniciusvarzea whom [proposed](https://github.com/ZiggyCreatures/FusionCache/issues/241) the `RecyclableMemoryStream` some versions ago, I'd like to see what they think of it.

Any idea?