@jodydonetti 2025-01-01
Hi @stebet , I just release [preview-4](https://github.com/ZiggyCreatures/FusionCache/releases/tag/v2.0.0-preview-4), which most probably will be the last preview before going GA with FusionCache V2 🥳

If you can play with it and let me know it would be great, thanks!